Interfaces complete

+++ src/Model/manager/PlayerManager.php
@@ -7,14 +7,17 @@ use PDO;
 use Model\Classes\Player;
 use DateTime;
 
-class PlayerManager {
+class PlayerManager implements ManagerInterface
+{
     private PDO $db;
 
-    public function __construct() {
+    public function __construct()
+    {
         $this->db = Database::getConnection();
     }
 
-    public function findAll(): array {
+    public function findAll(): array
+    {
 
         $stmt = $this->db->query("SELECT * FROM player");
 
@@ -33,12 +36,14 @@ class PlayerManager {
         return $players;
     }
 
-    public function getAge(Player $player): int {
+    public function getAge(Player $player): int
+    {
         $now = new DateTime();
         return $now->diff($player->getBirthdate())->y;
     }
 
-    public function findById(int $id): ?Player {
+    public function findById(int $id): ?Player
+    {
         $stmt = $this->db->prepare("SELECT * FROM player WHERE id = ?");
         $stmt->execute([$id]);
         $row = $stmt->fetch();
@@ -54,7 +59,13 @@ class PlayerManager {
         return null;
     }
 
-    public function insert(Player $player): bool {
+    public function insert(object $object): bool
+    {
+        if (!$object instanceof Player) {
+            return false;
+        }
+
+        $player = $object;
         $stmt = $this->db->prepare("INSERT INTO player (firstname, lastname, birthdate, picture) VALUES (?, ?, ?, ?)");
         return $stmt->execute([
             $player->getFirstname(),
@@ -64,12 +75,24 @@ class PlayerManager {
         ]);
     }
 
-    public function delete(Player $player): bool {
+    public function delete(object $object): bool
+    {
+        if (!$object instanceof Player) {
+            return false;
+        }
+
+        $player = $object;
         $stmt = $this->db->prepare("DELETE FROM player WHERE id = ?");
         return $stmt->execute([$player->getId()]);
     }
 
-    public function update(Player $player): bool {
+    public function update(object $object): bool
+    {
+        if (!$object instanceof Player) {
+            return false;
+        }
+
+        $player = $object;
         $stmt = $this->db->prepare("UPDATE player SET firstname = ?, lastname = ?, birthdate = ?, picture = ? WHERE id = ?");
         return $stmt->execute([
             $player->getFirstname(),
